add:selected card colors also seperated widgets
